move tests to LimitValidatorTest
