+++ src/test/java/net/sf/jsqlparser/util/validation/validator/LimitValidatorTest.java
@@ -9,14 +9,28 @@
  */
 package net.sf.jsqlparser.util.validation.validator;
 
+import java.util.Arrays;
 import org.junit.Test;
+import net.sf.jsqlparser.JSQLParserException;
 import net.sf.jsqlparser.util.validation.ValidationTestAsserts;
+import net.sf.jsqlparser.util.validation.feature.DatabaseType;
 
 public class LimitValidatorTest extends ValidationTestAsserts {
 
     @Test
-    public void test() {
-
+    public void testValidationLimitOffset() throws JSQLParserException {
+        String sql = "SELECT * FROM mytable WHERE mytable.col = 9 LIMIT 3, ?";
+        validateNoErrors(sql, 1, DatabaseType.MARIADB, DatabaseType.MYSQL);
     }
+
+    @Test
+    public void testValidationLimitAndOffset() throws JSQLParserException {
+        for (String sql : Arrays.asList("SELECT * FROM mytable WHERE mytable.col = 9 LIMIT 3",
+                "SELECT * FROM mytable WHERE mytable.col = 9 LIMIT ? OFFSET 3",
+                "SELECT * FROM mytable WHERE mytable.col = 9 OFFSET ?")) {
+            validateNoErrors(sql, 1, DatabaseType.MARIADB, DatabaseType.MYSQL, DatabaseType.POSTGRESQL);
+        }
+    }
+
 
 }
